test insertbank update bank get user

+++ models/UserModel.php
@@ -4,15 +4,27 @@ require_once 'BaseModel.php';
 
 class UserModel extends BaseModel {
 
-    public function findUserById($id) {
-        $sql = 'SELECT * FROM users WHERE id = '.$id;
+    protected static $_instance;
+
+    public function findUserById($id)
+    {
+        if(is_array($id) || is_object($id)){
+            return false;
+        }
+        $sql = 'SELECT * FROM users WHERE id = ' . $id;
         $user = $this->select($sql);
 
         return $user;
     }
+    
+    public function findUser($keyword)
+    {
+        if(is_array($keyword) || is_object($keyword)){
+            return false;
+        }
+        // $keyword = htmlentities($keyword, ENT_QUOTES, "UTF-8");
 
-    public function findUser($keyword) {
-        $sql = 'SELECT * FROM users WHERE user_name LIKE %'.$keyword.'%'. ' OR user_email LIKE %'.$keyword.'%';
+        $sql = 'SELECT * FROM users WHERE name LIKE %' . $keyword . '%' . ' OR user_email LIKE %' . $keyword . '%';
         $user = $this->select($sql);
 
         return $user;
@@ -25,11 +37,18 @@ class UserModel extends BaseModel {
      * @return array
      */
     public function auth($userName, $password) {
+        if(is_array($userName) || is_array($password) ||is_object($userName) || is_object($password)){
+            return false;
+        }
         $md5Password = md5($password);
-        $sql = 'SELECT * FROM users WHERE name = "' . $userName . '" AND password = "'.$md5Password.'"';
+        $sql = 'SELECT * FROM users WHERE name = "' . $userName . '" AND password = "' . $md5Password . '"';
 
         $user = $this->select($sql);
-        return $user;
+        //Check $user different is null?
+        if (!empty($user)) {
+            return true;
+        }
+        return false;
     }
 
     /**
@@ -38,6 +57,8 @@ class UserModel extends BaseModel {
      * @return mixed
      */
     public function deleteUserById($id) {
+        if (gettype($id) === "string") return "error";
+        if (!is_numeric($id)) return "error";
         $sql = 'DELETE FROM users WHERE id = '.$id;
         return $this->delete($sql);
 
@@ -49,8 +70,12 @@ class UserModel extends BaseModel {
      * @return mixed
      */
     public function updateUser($input) {
+
+        if(isset($input['name']) === null||isset($input['password']) === null||isset($input['email']) === null) return 'error!Không dc truyền đối tượng chuỗi';
+        if (!is_numeric($input['id'])||!is_int($input['id'])) return "error!Không dc truyền đối tượng chuỗi";
         $sql = 'UPDATE users SET 
                  name = "' . mysqli_real_escape_string(self::$_connection, $input['name']) .'", 
+                 email = "' . $input['email'] .'", 
                  password="'. md5($input['password']) .'"
                 WHERE id = ' . $input['id'];
 
@@ -65,8 +90,11 @@ class UserModel extends BaseModel {
      * @return mixed
      */
     public function insertUser($input) {
+        if(!is_string($input['name'])||!is_string($input['password'])||!is_string($input['email'])) return 'error!Không phải dữ liệu';
         $sql = "INSERT INTO `app_web1`.`users` (`name`, `password`) VALUES (" .
-                "'" . $input['name'] . "', '".md5($input['password'])."')";
+                "'" . $input['name'] . "', '".md5($input['password'])."', '"
+        . $input['email']
+        . "')";
 
         $user = $this->insert($sql);
 
@@ -80,12 +108,10 @@ class UserModel extends BaseModel {
      */
     public function getUsers($params = []) {
         //Keyword
+       
         if (!empty($params['keyword'])) {
             $sql = 'SELECT * FROM users WHERE name LIKE "%' . $params['keyword'] .'%"';
 
-            //Keep this line to use Sql Injection
-            //Don't change
-            //Example keyword: abcef%";TRUNCATE banks;##
             $users = self::$_connection->multi_query($sql);
         } else {
             $sql = 'SELECT * FROM users';
@@ -93,5 +119,14 @@ class UserModel extends BaseModel {
         }
 
         return $users;
+        
+    }
+
+    public static function getInstance() {
+        if (self::$_instance !== null){
+            return self::$_instance;
+        }
+        self::$_instance = new self();
+        return self::$_instance;
     }
 }
